electron/scripts/crop/python/requirements.txt
mediapipe>=0.10.9
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
